make autoMaterializeAssetEvaluationIds BigInteger instead of Ints (#25648)
BK

> Insert changelog entry or delete this section.

## Summary & Motivation

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.

+++ js_modules/dagster-ui/packages/ui-core/src/assets/AutoMaterializePolicyPage/PartitionSubsetListQuery.tsx
@@ -3,7 +3,7 @@ import {gql} from '../../apollo-client';
 export const PARTITION_SUBSET_LIST_QUERY = gql`
   query PartitionSubsetListQuery(
     $assetKey: AssetKeyInput!
-    $evaluationId: Int!
+    $evaluationId: BigInt!
     $nodeUniqueId: String!
   ) {
     truePartitionsForAutomationConditionEvaluationNode(
